requirements.txt
# Core SDK — brings in a2a-sdk, mcp, slim-bindings, and all transports
agntcy-app-sdk==0.5.5

# MCP protocol — includes FastMCP (mcp.server.fastmcp.FastMCP) and
# the MCP client (mcp.ClientSession, mcp.client.streamable_http).
# There is no separate "fastmcp" package; FastMCP is part of mcp itself.
mcp==1.27.0

# A2A protocol types (AgentCard, AgentSkill, …)
a2a-sdk==0.3.20
